+++ tests/conftest.py
@@ -5,6 +5,7 @@ import pytest
 from pycomprepair.core.plugin import PluginRegistry, reset_registry
 from pycomprepair.plugins.django_v5 import plugin as django_plugin
 from pycomprepair.plugins.fastapi_migration import plugin as fastapi_plugin
+from pycomprepair.plugins.numpy_v2 import plugin as numpy_plugin
 from pycomprepair.plugins.pydantic_v2 import plugin as pydantic_plugin
 from pycomprepair.plugins.sqlalchemy_v2 import plugin as sqlalchemy_plugin
 
@@ -31,6 +32,7 @@ def registry() -> PluginRegistry:
     reg.register(fastapi_plugin)
     reg.register(sqlalchemy_plugin)
     reg.register(django_plugin)
+    reg.register(numpy_plugin)
     return reg